OTEL: Fix #podNamePropagation test for Operator
Fixes #32490
